requirements.txt
asttokens==2.2.1
autopep8==2.0.1
backcall==0.2.0
colorama==0.4.6
comm==0.1.2
debugpy==1.6.4
decorator==5.1.1
distro==1.8.0
entrypoints==0.4
executing==1.2.0
ipykernel==6.19.4
ipython==8.7.0
jedi==0.18.2
jupyter_client==7.4.8
jupyter_core==5.1.2
llvmlite==0.39.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numba==0.56.4
numpy==1.23.5
packaging==22.0
pandas==1.5.2
parso==0.8.3
pdfminer==20191125
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.6.2
prompt-toolkit==3.0.36
psutil==5.9.4
pure-eval==0.2.2
pycodestyle==2.10.0
pycryptodome==3.16.0
Pygments==2.14.0
PyMuPDF==1.21.1
pypdf==3.2.0
python-dateutil==2.8.2
pytz==2022.7
pywin32==305
pyzmq==24.0.1
six==1.16.0
stack-data==0.6.2
tabula-py==2.6.0
tomli==2.0.1
tornado==6.2
traitlets==5.8.0
wcwidth==0.2.5
